+++ fathom-archetype-standard/src/main/resources/archetype-resources/src/main/java/conf/Caches.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package conf;
+package ${package}.conf;
 
 import fathom.conf.Settings;
 import fathom.jcache.CachesModule;