+++ fathom-archetype-standard/src/main/resources/archetype-resources/src/main/java/conf/Jobs.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package conf;
+package ${package}.conf;
 
 import fathom.quartz.JobsModule;
 import fathom.quartz.Scheduled;