+++ fathom-archetype-standard/src/main/resources/archetype-resources/src/main/java/controllers/SecureApiController.java
@@ -14,11 +14,13 @@
  * limitations under the License.
  */
 
-package controllers;
+package ${package}.controllers;
+
+import ${package}.dao.ItemDao;
+import ${package}.models.Item;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
-import dao.ItemDao;
 import fathom.metrics.Metered;
 import fathom.realm.Account;
 import fathom.rest.controller.Auth;
@@ -27,7 +29,6 @@ import fathom.rest.controller.GET;
 import fathom.rest.controller.Path;
 import fathom.rest.controller.Produces;
 import fathom.rest.security.aop.RequirePermission;
-import models.Item;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;