+++ fathom-archetype-standard/src/main/resources/archetype-resources/src/main/java/models/Item.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package models;
+package ${package}.models;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;