+++ fathom-archetype-standard/src/main/resources/archetype-resources/src/test/java/conf/RoutesTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package conf;
+package ${package}.conf;
 
 import fathom.test.FathomTest;
 import org.junit.Test;