+++ fathom-archetype-standard/src/main/resources/archetype-resources/src/test/java/controllers/ApiControllerTest.java
@@ -1,4 +1,4 @@
-package controllers;
+package ${package}.controllers;
 
 import fathom.test.FathomTest;
 import org.junit.Test;